Split serializers and views into dedicated files
