+++ metrics/api/views/__init__.py
@@ -1,0 +1,5 @@
+from .charts import ChartsView, ChartView
+from .file_upload import FileUploadView
+from .headlines import HeadlinesView
+from .health import HealthView
+from .trends import TrendsView
