+++ metrics/api/views/charts.py
@@ -1,41 +1,14 @@
 import os
 from http import HTTPStatus
 
-from django.http import FileResponse, HttpResponse
+from django.http import FileResponse
 from drf_spectacular.utils import extend_schema
-from rest_framework.parsers import MultiPartParser
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework_api_key.permissions import HasAPIKey
 
-from metrics.api.serializers import (
-    ChartsQuerySerializer,
-    ChartsSerializer,
-    HeadlinesQuerySerializer,
-    TrendsQuerySerializer,
-    TrendsResponseSerializer,
-)
-from metrics.data.operations.api_models import generate_api_time_series
-from metrics.data.operations.core_models import load_core_data
-from metrics.domain.trends.state import TREND_AS_DICT
+from metrics.api.serializers import ChartsQuerySerializer, ChartsSerializer
 from metrics.interfaces.charts import access, data_visualization_superseded, validation
-from metrics.interfaces.headlines.access import (
-    BaseInvalidHeadlinesRequestError,
-    generate_headline_number,
-)
-from metrics.interfaces.trends.access import (
-    TrendNumberDataNotFoundError,
-    generate_trend_numbers,
-)
-
-
-class HealthView(APIView):
-    permission_classes = []
-
-    @staticmethod
-    def get(*args, **kwargs):
-        return HttpResponse(HTTPStatus.OK.value)
-
 
 DEPRECATION_DATE_CHARTS_ENDPOINT = "Wed, 19 Apr 2023 23:59:59 GMT"
 DEPRECATION_MESSAGE_CHARTS = f"This endpoint has been deprecated. This functionality can now be found within the `/charts/v2/` endpoint. Deprecation date: {DEPRECATION_DATE_CHARTS_ENDPOINT}"
@@ -102,29 +75,6 @@ class ChartView(APIView):
         os.remove(filename)
 
         return response
-
-
-class FileUploadView(APIView):
-    parser_classes = [MultiPartParser]
-    permission_classes = [HasAPIKey]
-
-    @extend_schema(
-        operation_id="upload_file",
-        request={
-            "multipart/form-data": {
-                "type": "object",
-                "properties": {"file": {"type": "string", "format": "binary"}},
-            }
-        },
-        deprecated=True,
-    )
-    def put(self, request, *args, **kwargs):
-        """
-        Note that this endpoint is **deprecated** and should only be used for demo/testing purposes.
-        """
-        load_core_data(filename=request.FILES.get("file"))
-        generate_api_time_series()
-        return Response(status=204)
 
 
 class ChartsView(APIView):
@@ -227,118 +177,3 @@ class ChartsView(APIView):
         os.remove(filename)
 
         return response
-
-
-class HeadlinesView(APIView):
-    permission_classes = [HasAPIKey]
-
-    @extend_schema(parameters=[HeadlinesQuerySerializer])
-    def get(self, request, *args, **kwargs):
-        """This endpoint can be used to retrieve headline-type numbers for a given `metric` & `topic` combination.
-
-        ---
-
-        # Main errors
-
-        Note that this endpoint will only return single-headline number type data.
-        If the `metric` provided relates to timeseries type data then the request will be deemed invalid.
-
-        ---
-
-        ### Timeseries type metrics are invalid
-
-        For example, a request for the following would be **invalid**:
-
-        - metric =`new_cases_daily`
-
-        - topic = `COVID-19`
-
-        This would be **invalid** because the `metric` of `new_cases_daily` relates to timeseries data,
-        which is not represented by a single headline-type figure.
-
-        ---
-
-        Whereas, a request for the following would be **valid**:
-
-        - metric =`new_cases_7days_sum`
-
-        - topic = `COVID-19`
-
-        This would be **valid** because the `metric` of `new_cases_7days_sum` relates to headline data,
-        which can be represented by a single headline-type figure.
-
-        """
-        query_serializer = HeadlinesQuerySerializer(data=request.query_params)
-        query_serializer.is_valid(raise_exception=True)
-
-        topic = query_serializer.data["topic"]
-        metric = query_serializer.data["metric"]
-
-        try:
-            headline_number: str = generate_headline_number(topic=topic, metric=metric)
-        except BaseInvalidHeadlinesRequestError as error:
-            return Response(
-                status=HTTPStatus.BAD_REQUEST, data={"error_message": str(error)}
-            )
-
-        return Response({"value": headline_number})
-
-
-class TrendsView(APIView):
-    permission_classes = [HasAPIKey]
-
-    @extend_schema(
-        parameters=[TrendsQuerySerializer],
-        responses={HTTPStatus.OK: TrendsResponseSerializer},
-    )
-    def get(self, request, *args, **kwargs):
-        """This endpoint can be used to retrieve trend-type data for a given `topic`, `metric` and `percentage_metric` combination.
-
-        The response will include data to indicate whether the change should be considered positive.
-
-        Primarily, this is because an increase in metric value requires additional context whether it should be considered positive.
-
-        For example, an increase in cases would be considered negative.
-        But an increase in vaccinations would be considered positive.
-
-        ---
-
-        # Main errors
-
-        Note that this endpoint will only key-value pair type data for the trend block.
-        If the provided parameters return data which does not exist then this will be invalid.
-
-        ---
-
-        ## Data could not be found
-
-        For example, a request for the following would be **invalid**:
-
-        - topic = `COVID-19`
-
-        - metric =`weekly_positivity`
-
-        - percentage_metric =`new_deaths_7days_change_percentage`
-
-        This would be **invalid** because the `metric` of `weekly_positivity` `COVID-19` will not return any data.
-
-        """
-        query_serializer = TrendsQuerySerializer(data=request.query_params)
-        query_serializer.is_valid(raise_exception=True)
-
-        topic: str = query_serializer.data["topic"]
-        metric_name: str = query_serializer.data["metric"]
-        percentage_metric_name: str = query_serializer.data["percentage_metric"]
-
-        try:
-            trends_data: TREND_AS_DICT = generate_trend_numbers(
-                topic=topic,
-                metric_name=metric_name,
-                percentage_metric_name=percentage_metric_name,
-            )
-        except TrendNumberDataNotFoundError as error:
-            return Response(
-                status=HTTPStatus.BAD_REQUEST, data={"error_message": str(error)}
-            )
-
-        return Response(trends_data)
