+++ metrics/api/views/trends.py
@@ -1,0 +1,73 @@
+from http import HTTPStatus
+
+from drf_spectacular.utils import extend_schema
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework_api_key.permissions import HasAPIKey
+
+from metrics.api.serializers import TrendsQuerySerializer, TrendsResponseSerializer
+from metrics.domain.trends.state import TREND_AS_DICT
+from metrics.interfaces.trends.access import (
+    TrendNumberDataNotFoundError,
+    generate_trend_numbers,
+)
+
+
+class TrendsView(APIView):
+    permission_classes = [HasAPIKey]
+
+    @extend_schema(
+        parameters=[TrendsQuerySerializer],
+        responses={HTTPStatus.OK: TrendsResponseSerializer},
+    )
+    def get(self, request, *args, **kwargs):
+        """This endpoint can be used to retrieve trend-type data for a given `topic`, `metric` and `percentage_metric` combination.
+
+        The response will include data to indicate whether the change should be considered positive.
+
+        Primarily, this is because an increase in metric value requires additional context whether it should be considered positive.
+
+        For example, an increase in cases would be considered negative.
+        But an increase in vaccinations would be considered positive.
+
+        ---
+
+        # Main errors
+
+        Note that this endpoint will only key-value pair type data for the trend block.
+        If the provided parameters return data which does not exist then this will be invalid.
+
+        ---
+
+        ## Data could not be found
+
+        For example, a request for the following would be **invalid**:
+
+        - topic = `COVID-19`
+
+        - metric =`weekly_positivity`
+
+        - percentage_metric =`new_deaths_7days_change_percentage`
+
+        This would be **invalid** because the `metric` of `weekly_positivity` `COVID-19` will not return any data.
+
+        """
+        query_serializer = TrendsQuerySerializer(data=request.query_params)
+        query_serializer.is_valid(raise_exception=True)
+
+        topic: str = query_serializer.data["topic"]
+        metric_name: str = query_serializer.data["metric"]
+        percentage_metric_name: str = query_serializer.data["percentage_metric"]
+
+        try:
+            trends_data: TREND_AS_DICT = generate_trend_numbers(
+                topic=topic,
+                metric_name=metric_name,
+                percentage_metric_name=percentage_metric_name,
+            )
+        except TrendNumberDataNotFoundError as error:
+            return Response(
+                status=HTTPStatus.BAD_REQUEST, data={"error_message": str(error)}
+            )
+
+        return Response(trends_data)
